refactor(antigravity): extract AntigravitySharedCommand base (review #1)
Address code-review finding #1 on PR #1708: AntigravityIdeCommand was a
near-verbatim copy of the legacy AntigravityCommand, inconsistent with the
MCP/Skills features in this PR which factored shared logic into
AntigravityMcp / AntigravitySharedSkill base classes.

Introduce AntigravitySharedCommand holding the shared workflow body/trigger
transform, frontmatter validation, and round-trip logic. AntigravityIdeCommand
becomes a thin subclass overriding only the project/global workflows
directories and its rulesync target name. The legacy antigravity alias keeps
its own standalone class (mirroring the AntigravitySkill precedent for the
deprecated singular .agent/ tree). No behavior change.
