+++ src/features/commands/antigravity-ide-command.ts
@@ -1,246 +1,33 @@
-import { basename, join } from "node:path";
+import { join } from "node:path";
 
-import { AiFileParams, ValidationResult } from "../../types/ai-file.js";
-import { formatError } from "../../utils/error.js";
-import { readFileContent } from "../../utils/file.js";
-import { parseFrontmatter, stringifyFrontmatter } from "../../utils/frontmatter.js";
-import { isRecord } from "../../utils/type-guards.js";
-import {
-  AntigravityCommandFrontmatter,
-  AntigravityCommandFrontmatterSchema,
-} from "./antigravity-command.js";
-import { RulesyncCommand, RulesyncCommandFrontmatter } from "./rulesync-command.js";
-import {
-  ToolCommand,
-  ToolCommandForDeletionParams,
-  ToolCommandFromFileParams,
-  ToolCommandFromRulesyncCommandParams,
-  ToolCommandSettablePaths,
-} from "./tool-command.js";
-
-export type AntigravityIdeCommandParams = {
-  frontmatter: AntigravityCommandFrontmatter;
-  body: string;
-} & AiFileParams;
+import { ToolTarget } from "../../types/tool-targets.js";
+import { AntigravitySharedCommand } from "./antigravity-shared-command.js";
+import { RulesyncCommand } from "./rulesync-command.js";
 
 /**
  * Command (workflow) generator for the Google Antigravity IDE (Antigravity 2.0).
  *
- * Generates workflow files for Antigravity's `.agents/workflows/` directory
- * (project scope) and `~/.gemini/antigravity/global_workflows/` (global scope).
- * The workflow body/frontmatter shape matches {@link AntigravityCommand}.
+ * Generates workflow files in `.agents/workflows/` (project scope) and
+ * `~/.gemini/antigravity/global_workflows/` (global scope). All body and
+ * frontmatter handling is shared with {@link AntigravitySharedCommand}.
  */
-export class AntigravityIdeCommand extends ToolCommand {
-  private readonly frontmatter: AntigravityCommandFrontmatter;
-  private readonly body: string;
-
-  static getSettablePaths({ global = false }: { global?: boolean } = {}): ToolCommandSettablePaths {
-    if (global) {
-      return {
-        relativeDirPath: join(".gemini", "antigravity", "global_workflows"),
-      };
-    }
-    return {
-      relativeDirPath: join(".agents", "workflows"),
-    };
+export class AntigravityIdeCommand extends AntigravitySharedCommand {
+  protected static override getProjectRelativeDirPath(): string {
+    return join(".agents", "workflows");
   }
 
-  constructor({ frontmatter, body, ...rest }: AntigravityIdeCommandParams) {
-    if (rest.validate) {
-      const result = AntigravityCommandFrontmatterSchema.safeParse(frontmatter);
-      if (!result.success) {
-        throw new Error(
-          `Invalid frontmatter in ${join(rest.relativeDirPath, rest.relativeFilePath)}: ${formatError(result.error)}`,
-        );
-      }
-    }
-
-    super({
-      ...rest,
-      fileContent: stringifyFrontmatter(body, frontmatter),
-    });
-
-    this.frontmatter = frontmatter;
-    this.body = body;
+  protected static override getGlobalRelativeDirPath(): string {
+    return join(".gemini", "antigravity", "global_workflows");
   }
 
-  getBody(): string {
-    return this.body;
+  protected override getToolTargetName(): ToolTarget {
+    return "antigravity-ide";
   }
 
-  getFrontmatter(): Record<string, unknown> {
-    return this.frontmatter;
-  }
-
-  toRulesyncCommand(): RulesyncCommand {
-    const { description, ...restFields } = this.frontmatter;
-
-    const rulesyncFrontmatter: RulesyncCommandFrontmatter = {
-      targets: ["antigravity-ide"],
-      description,
-      ...(Object.keys(restFields).length > 0 && { antigravity: restFields }),
-    };
-
-    const fileContent = stringifyFrontmatter(this.body, rulesyncFrontmatter);
-
-    return new RulesyncCommand({
-      outputRoot: ".",
-      frontmatter: rulesyncFrontmatter,
-      body: this.body,
-      relativeDirPath: RulesyncCommand.getSettablePaths().relativeDirPath,
-      relativeFilePath: this.relativeFilePath,
-      fileContent,
-      validate: true,
-    });
-  }
-
-  private static extractAntigravityConfig(
-    rulesyncCommand: RulesyncCommand,
-  ): Record<string, unknown> | undefined {
-    const antigravity = rulesyncCommand.getFrontmatter().antigravity;
-    return isRecord(antigravity) ? antigravity : undefined;
-  }
-
-  private static resolveTrigger(
-    rulesyncCommand: RulesyncCommand,
-    antigravityConfig: Record<string, unknown> | undefined,
-  ): string {
-    const rulesyncFrontmatter = rulesyncCommand.getFrontmatter();
-
-    const antigravityTrigger =
-      antigravityConfig && typeof antigravityConfig.trigger === "string"
-        ? antigravityConfig.trigger
-        : undefined;
-
-    const rootTrigger =
-      typeof rulesyncFrontmatter.trigger === "string" ? rulesyncFrontmatter.trigger : undefined;
-
-    const bodyTriggerMatch = rulesyncCommand.getBody().match(/trigger:\s*(\/[\w-]+)/);
-
-    const filenameTrigger = `/${basename(rulesyncCommand.getRelativeFilePath(), ".md")}`;
-
-    return (
-      antigravityTrigger ||
-      rootTrigger ||
-      (bodyTriggerMatch ? bodyTriggerMatch[1] : undefined) ||
-      filenameTrigger
-    );
-  }
-
-  static fromRulesyncCommand({
-    outputRoot = process.cwd(),
-    rulesyncCommand,
-    validate = true,
-    global = false,
-  }: ToolCommandFromRulesyncCommandParams): AntigravityIdeCommand {
-    const rulesyncFrontmatter = rulesyncCommand.getFrontmatter();
-    const antigravityConfig = this.extractAntigravityConfig(rulesyncCommand);
-
-    const trigger = this.resolveTrigger(rulesyncCommand, antigravityConfig);
-
-    const turbo = typeof antigravityConfig?.turbo === "boolean" ? antigravityConfig.turbo : true;
-
-    let body = rulesyncCommand
-      .getBody()
-      .replace(/^---\r?\n[\s\S]*?\r?\n---\r?\n/, "")
-      .trim();
-
-    // Sanitize trigger to prevent path traversal (e.g. /../evil).
-    const sanitizedTrigger = trigger.replace(/[^a-zA-Z0-9-_]/g, "-").replace(/^-+|-+$/g, "");
-    if (!sanitizedTrigger) {
-      throw new Error(`Invalid trigger: sanitization resulted in empty string from "${trigger}"`);
-    }
-    const relativeFilePath = `${sanitizedTrigger}.md`;
-
-    const turboDirective = turbo ? "\n\n// turbo" : "";
-    body = `# Workflow: ${trigger}\n\n${body}${turboDirective}`;
-
-    const description = rulesyncFrontmatter.description;
-
-    const antigravityFrontmatter: AntigravityCommandFrontmatter = {
-      description,
-      trigger,
-      turbo,
-    };
-
-    const fileContent = stringifyFrontmatter(body, antigravityFrontmatter);
-    const paths = this.getSettablePaths({ global });
-
-    return new AntigravityIdeCommand({
-      outputRoot,
-      frontmatter: antigravityFrontmatter,
-      body,
-      relativeDirPath: paths.relativeDirPath,
-      relativeFilePath,
-      fileContent,
-      validate,
-    });
-  }
-
-  validate(): ValidationResult {
-    if (!this.frontmatter) {
-      return { success: true, error: null };
-    }
-
-    const result = AntigravityCommandFrontmatterSchema.safeParse(this.frontmatter);
-    if (result.success) {
-      return { success: true, error: null };
-    }
-    return {
-      success: false,
-      error: new Error(
-        `Invalid frontmatter in ${join(this.relativeDirPath, this.relativeFilePath)}: ${formatError(result.error)}`,
-      ),
-    };
-  }
-
-  static isTargetedByRulesyncCommand(rulesyncCommand: RulesyncCommand): boolean {
+  static override isTargetedByRulesyncCommand(rulesyncCommand: RulesyncCommand): boolean {
     return this.isTargetedByRulesyncCommandDefault({
       rulesyncCommand,
       toolTarget: "antigravity-ide",
-    });
-  }
-
-  static async fromFile({
-    outputRoot = process.cwd(),
-    relativeFilePath,
-    validate = true,
-    global = false,
-  }: ToolCommandFromFileParams): Promise<AntigravityIdeCommand> {
-    const paths = this.getSettablePaths({ global });
-    const filePath = join(outputRoot, paths.relativeDirPath, relativeFilePath);
-    const fileContent = await readFileContent(filePath);
-    const { frontmatter, body: content } = parseFrontmatter(fileContent, filePath);
-
-    const result = AntigravityCommandFrontmatterSchema.safeParse(frontmatter);
-    if (!result.success) {
-      throw new Error(`Invalid frontmatter in ${filePath}: ${formatError(result.error)}`);
-    }
-
-    return new AntigravityIdeCommand({
-      outputRoot,
-      relativeDirPath: paths.relativeDirPath,
-      relativeFilePath,
-      frontmatter: result.data,
-      body: content.trim(),
-      fileContent,
-      validate,
-    });
-  }
-
-  static forDeletion({
-    outputRoot = process.cwd(),
-    relativeDirPath,
-    relativeFilePath,
-  }: ToolCommandForDeletionParams): AntigravityIdeCommand {
-    return new AntigravityIdeCommand({
-      outputRoot,
-      relativeDirPath,
-      relativeFilePath,
-      frontmatter: { description: "" },
-      body: "",
-      fileContent: "",
-      validate: false,
     });
   }
 }
